resize image automatically when added

+++ Backend/controller/DeckController.js
@@ -65,7 +65,6 @@ module.exports = {
 
         Deck.find({ $or: [{ owner: req.payload._id }, { subscribers: { $eq: req.payload._id } }] }).populate('owner', 'email').populate('cards').exec().then(decks => {
             if (decks) {
-                console.log(decks);
                 let aDecks = new Array();
                 decks.forEach(deck => {
                     let d = {
@@ -116,7 +115,7 @@ module.exports = {
             deck.save().then(deck => {
                 user.decks.push(deck);
                 user.save().then(user => {
-                    res.status(200).json({_id:deck.id});
+                    res.status(200).json({ _id: deck.id });
                 }, error => {
                     console.log(error);
                     res.status(500).send(error);
@@ -144,15 +143,26 @@ module.exports = {
                 deck.averageRating = req.body.deck.averageRating;
                 deck.isPublic = req.body.deck.isPublic;
 
-                req.body.deck.cards.forEach( async card => {
-                    if (card._id) {
-                        await Card.findOneAndUpdate({ _id: card._id }, card, { new: true });
-                    } else {
-                        let c = new Card({ question: card.question, answer: card.answer, difficulty: card.difficulty });
-                        c.save();
-                        deck.cards.push(c);
-                    }
-                });
+                if(req.body.deck.cards.length>0){
+                    req.body.deck.cards.forEach(async card => {
+                        if (card._id) {
+                            await Card.findOneAndUpdate({ _id: card._id }, card, { new: true });
+                        } else {
+                            let c = new Card({ question: card.question, answer: card.answer, difficulty: card.difficulty });
+                            c.save();
+                            deck.cards.push(c);
+                        }
+                    });
+                } else {
+                    Card.find({ _id: { $in: deck.cards } }).then(cards => {
+                        cards.forEach(card => {
+                            card.remove();
+                        });
+                    }).catch(error => {
+                        console.log(error);
+                    })
+                    // deck.cards = req.body.deck.cards;
+                }
 
                 deck.save().then(savedDeck => {
                     res.status(200).send();
@@ -165,26 +175,23 @@ module.exports = {
             console.log(error);
         });
     },
-    deleteDeck: function (req, res) { 
+    deleteDeck: function (req, res) {
         let deckId = req.query.id;
 
-        // TODO: only delete if owner !!!
-
-
-        Deck.findOneAndDelete({_id: deckId}).then(deck => {
-            if(deck) {
-                Card.find({_id: {$in: deck.cards}}).then(cards=> {
+        Deck.findOneAndDelete({ $and: [{ _id: deckId, owner: req.payload._id }] }).then(deck => {
+            if (deck) {
+                Card.find({ _id: { $in: deck.cards } }).then(cards => {
                     cards.forEach(card => {
                         card.remove();
                     });
                 }).catch(error => {
                     console.log(error);
                 })
-                User.find({_id: {$in: deck.subscribers}}).then(users => {
+                User.find({ _id: { $in: deck.subscribers } }).then(users => {
                     users.forEach(user => {
-                        for(var i = 0; i < user.decks.length; i++){
-                            if(user.decks[i]._id.equals(deck._id)){
-                                user.decks.splice(i,1)
+                        for (var i = 0; i < user.decks.length; i++) {
+                            if (user.decks[i]._id.equals(deck._id)) {
+                                user.decks.splice(i, 1)
                                 i--;
                             }
                         }
@@ -198,5 +205,13 @@ module.exports = {
                 res.status(401).send();
             }
         })
+    },
+
+    subscribeDeck: function (req, res) {
+        
+    },
+
+    unsubscribeDeck: function (req, res) {
+
     }
 }
